Refactoring #2
