+++ src/Types/Inline/QueryResult/AbstractInlineQueryResult.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace TelegramBot\Api\Types\Inline;
+namespace TelegramBot\Api\Types\Inline\QueryResult;
 
 use TelegramBot\Api\BaseType;
